==> nb0-1_create_dataset.ipynb <==
# Intro

For generating dataset of shape (n_samples, max_seq_len, feats_per_step) from the original dataframe 

This version is the default

import numpy as np
import pandas as pd
from matplotlib import pyplot as plt
import os

import pickle
from datetime import date

from my_lib.encoding import *

RAW_CSV_DIR = "raw_czech"

for fname in os.listdir(RAW_CSV_DIR):
    if ".csv" in fname and "new_" in fname:
        print(fname)
        csv_name = fname.split(".")[0]
        cmd = f"{csv_name} = pd.read_csv('{RAW_CSV_DIR}/{fname}')"
        print("Running:", cmd)
        exec(cmd)


### Setup

max_seq_len = 80
min_seq_len = 20

### Sort by acct 

tr_by_acct = new_transaction.sort_values(by = ["account_id", "date"])

diffs = np.diff(tr_by_acct["account_id"]) > 0   # detect acct change
np.where(diffs)

tr_by_acct

new_client[tr_by_acct["account_id"]]

new_client

len(tr_by_acct), len(new_disposition)

ud = new_disposition["account_id"].unique()
ut = tr_by_acct["account_id"].unique()

len(ud), len(ut)

sum(ud-ut)

from datetime import datetime
czech_date_parser = lambda x: datetime.strptime(str(x), "%y%m%d")


df = pd.merge(
        pd.merge(
            tr_by_acct,
            new_disposition[new_disposition["type"]=="OWNER"][["account_id", "client_id"]],
            on="account_id",
        ),
        new_client[["client_id", "age"]],
        on = "client_id"
)
df

df["datetime"] = df["date"].apply(czech_date_parser)
df

df.to_csv("data/tr_by_acct_w_age.csv", index=False)
