==> nb0-2_clear_old.ipynb <==
!rm -r checkpoints/

!rm -r raw_czech/

!rm -r stored_data/

!rm -r generated_data/

!rm -r training_history/

!mkdir checkpoints/

!mkdir stored_data/

!mkdir generated_data/

!mkdir training_history/

!mkdir raw_czech/
